internals - rename internal RNG support funcs and vars
This makes it so that the internal RNG support functions
and variables are named the same as their non-internal
equivalents, but with _internal as a suffix. This should
make it easier to find them. For instance PL_internal_random_state
will not show up when searching for PL_random_state, but
PL_random_state_internal will.