configure should not do anything about random number generators
We provide our own RNG, we do not need any information from Configure
about the system RNG. Configure basically should not be involved in such
things, just like it isn't with regard to hash functions.

This is a placeholder commit so that Tux knows what need to be done and
for discussion purposes. It isn't intended for direct merging. The
change needs to be done the normal way for Configure changes.