regen/embed.pl - handle conditional duplicate symbols properly
If we have a symbol which has a different definition depending on a c
pre-processor condition, include the condition when de-duping.

For instance if in intrpvar.h we had

    #ifdef HAS_QUAD
    PERLVAR(I, SomeField, U64TYPE)
    #else
    PERLVAR(I, SomeField, U32)
    #endif

We don't want to throw warnings about duplicate SomeField, as only one
of the definitions will be active at a time.

This includes changes to regen/HeaderParser.pl to facilitate this, by
adding a cond_as_str() method to the HeaderLine objects.