op.c: call class_wrap_method_body() for lexical methods
Class method bodies must be wrapped by this function, which adds the
OP_METHSTART op at the beginning, which is responsible for shifting
the `$self` lexical from the arguments array, and setting up field
bindings. This was already done for non-lexical (i.e. package) methods
but was forgotten for lexical methods, so they didn't work properly,
leading to the bug found in GH#23030 whereby the argument count in a
signatured lexical method did not take account of the `$self` argument
properly.

In addition, fields accessed by the method body would not have been
working, though originally this wasn't reported or tested either. A test
for this has also been added.