cflags.SH: don't try to add both -Wextra and -W
-W is just the older spelling of -Wextra, so don't try to add both.
Also, remove the special-case logic for -W that skips adding it if
-Wextra was already added.

Background: -W was officially renamed to -Wextra in gcc 3.4
<https://gcc.gnu.org/gcc-3.4/changes.html>, released in April 2004. So
the only case where this code had any effect was 20+ years old releases
of gcc that understood -W, but not -Wextra. I don't think we need to
cater to those compilers for modern perl development.