ParseXS: refactor: add multiline,code,PREINIT node
Add these Node subclasses:

    ExtUtils::ParseXS::Node::multiline
    ExtUtils::ParseXS::Node::code
    ExtUtils::ParseXS::Node::PREINIT

The first is a very generic base class for the many planned node types
which will represent multi-line XS keywords, such as

    FOO: aaa
       bbb
       ccc

The lines are just read into an array ref.

ExtUtils::ParseXS::Node::code is a generic subclass of Node::multiline
which represents keywords that contain sections of C code, such as
PREINIT, PPCODE etc. It does extra processing such as skipping leading
blank lines and wrapping the output in '#line ...'.

ExtUtils::ParseXS::Node::PREINIT is a concrete subclass of Node::code
which represents the PREINIT keyword. This is kind of a
proof-of-concept; other such code keywords (CODE, PPCODE etc) will be
added later.

The net effect of this commit is to move processing of the PREINIT:
keyword into the parse() and as_code() methods of the Node::PREINIT
class (and/or its parents) and away from the print_section() and
PREINIT_handler() methods in ExtUtils::ParseXS.

This is intended as a small incremental step towards having a real AST.

A PREINIT object is currently created, parsed, printed and destroyed all
at the same point in time. In some future refactoring, the intention is
that the object will be stored in a tree, and the parsing and
code-emitting steps will be done at different times.