pp_ord: Use utf8_to_uv over utf8n_to_uvchr
This is the more modern spelling.

This also changes the code so that 0 is always explicitly returned if
the SV is zero length.  Previously it would only do this if the SV was
marked as being in UTF-8.  If it wasn't, it would return the first byte
in the PV.  That should also always be a NUL or 0, but now it is
explicit.