regexec.c: Use utf8_to_uv_or_die over utf8n_to_uvchr
This will now die if the input is malformed.  We decided for 5.42 that
it is pointless to continue on in pattern matching in the face of
malformed input.  This was the final occurrence of a to_uvchr() function
in this file; the rest were converted for 5.42.  This was deferred
because of the lateness of the development cycle, and it didn't fix an
obvious bug