pp_pack: NEXT_UNI_VAL: Use 'uv' form, not 'uvchr'
This converts a call to utf8n_to_uvchr() to instead use
utf8_to_uv_flags().

This is part of the process of replacing all the '_uvchr' functions with
'_uv' functions, which are now preferred.  See perlapi.

There is a subtle change of behavior here when the string to convert is
malformed.  I believe this is a bug fix.  The macro NEXT_UNI_VAL assumed
the failure return of utf8n_to_uvchr() is 'retlen' being -1.  However,
that value is only ever returned if the flags passed to the function
include UTF8_CHECK_ONLY.  Inspection of the code showed that this macro
is never called with that flag; so this never currently fails.

I looked at Perl v5.8.9, and the situation was the same.  I doubt it has
changed until now.

This commit changes to use utf8_to_uv_flags(), and its failure return is
properly checked for.  The upshot is that this could now fail; whereas
it couldn't before, even though the clear intent of the code has been to
fail upon error.