utf8_to_uv_msgs: Simplfiy checking for overlong
Prior to this commit, there were two different methods for doing this
check; one if no malformations have been found so far, and the other if
some had been found.  The latter method is valid in both cases, and is
just as fast or faster than the first method.  So change to always use
it