utf8_to_uv_msgs: Make warning message consistent
This inserts a colon in the message to make it consistent with all the
other similar ones.  This will enable easier parsing of messages for
future commits.

If this change causes failures in any code in the field, that would be a
Good Thing, as that code is now illegal.  This message happens only when
this function is called with a zero length string, which has been
illegal for some releases.  Doing this in a DEBUGGING build results in
an assert(0).  This message now only exists as a fallback to prevent
bad or undefined behavior, should we get called with that anyway.