Perl_utf8_to_uv_msgs_helper: Use bit pos in switch()
The previous commit #defined the bit position of various flag bits.
This commit changes to use those as cases in a switch(), rather than the
shifted values.  This results in better generated code as the cases
now reduce to:             case 0:     case 1:     case 2:    ...
as opposed to previously:  case 1<<0:  case 1<<1:  case 1<<2: ...

clang and gcc were unable to see the pattern previously but now are.