utf8_to_uv_msgs: Revise and rename macro
This macro is used to hide the details of determining if an abnormal
condition should raise a warning or not.  But I found it more convenient
to expand the macro to return the packed warnings category(ies) if a
warning should be raised or not.  That information is known inside the
macro and was being discarded, and then having to be recalculated.  The
new name reflects its expanded purpose, PACK_WARN.  0 is returned if no
warnings need be raised; and importantly fixing a bug in the old code,
it returns < 0 if no warning should be raised directly, but that an
entry needs to be added to the AV array returned by the function (if the
parameter requesting that has been passed in)

But Encode, for which this form of the translation function was created,
and may be the only user of it, depends on not getting a zero return.
So this has an override until Encode can be fixed.

I introduced the DIE_IF_MALFORMED flag in the previous development
release, making it subservient to the CHECK_ONLY flag.  I have since
realized that the precedence should be reversed.  If a developer
inadvertently passes both flags, it is better to honor the one saying
you need to quit, than the one saying ignore any problems.