utf8_to_uv_msgs: Reverse order of finding overflow/extended UTF-8
This begins the process of fixing the current problematic behavior of
handling UTF-8 that is for code points above the Unicode maximum.

The lowest of these are considered SUPERs, but if you go high enough, it
takes Perl's extended UTF-8 to represent them.  Higher still, and the
extended UTF-8 can represent code points that don't fit in the current
platform's word size.

A complication is overlongs, where the representation for a seemingly
large code point can reduce down to something much smaller; even 0.
Such sequences are considered invalid by fiat from Unicode due to
successful hacker attacks using them.  But Perl has traditionally
allowed XS code to allow them, with flags passed to the translation
functions.  So it is important to get it right.

A sequence that overflows by necessity is using Perl's extended UTF-8,
as that kicks in below a 32 bit word.  This commit reverses the prior
order of testing for overflow and extended UTF-8.  Steps can be saved
because we now test for Perl-extended first, which is a lot more likely
to happen than overflow.