utf8_to_uv_msgs: Fix handling of too-short malformations
At this point in the code we know that the input sequence is shorter
than a full character and that it is the legal beginning of a sequence that
could evaluate to a code point that is of interest to the caller of this
function.  It turns out that in some cases any filling out of the input
to a full character must lead to a code point that the caller is
interested in.  That interest has been signalled by flags passed to this
function.

In the past, we filled out the sequence with the minimum legal
continuation byte, but that is wrong for some cases.  This commit fixes
that.

Certain start bytes require the second byte to be higher than the
minimum, or else it is an overlong.  Prior to this commit, we could
generate overlongs.  This commit avoids that pitfall.

It also moves the complex analysis away from the comments in the code,
and to this commit message, adding even more analysis.

There are four classes of code points that the caller can have signalled
to this function that it is interested in.

The noncharacter code point class always needs a full sequence to
determine, and the conditionals prevent the code this analasys is about
from being executed.

Use of Perl extended-UTF-8 is determinable from the first byte in the
input sequence, and that has already been determined.

Both of the other two sequences don't have to be fully filled out in
order to determine if a partial sequence would lead to them or not.

Consider first, the sequences that evaluate to an above-Unicode code
point, charmingly named "supers" by Perl's poetic coders.
                ASCII platforms          EBCDIC I8
     U+10FFFF: \xF4\x8F\xBF\xBF    \xF9\xA1\xBF\xBF\xBF
     0x110000: \xF4\x90\x80\x80    \xF9\xA2\xA0\xA0\xA0
                 *
(Continuation byte range):
                \x80 to \xbf           \xa0 to \xbf

On ASCII platforms, any start byte \xf3 and below can't be for a super,
and any non-overlong sequence \xf5 and above has to be for a super.  If
the start byte is \xf4, we need a second byte to resolve the ambiguity.
But it takes just the one, or possibly two bytes to make the
determination.  It's similar on EBCDIC, but with different values.

And a similar situation exists for the surrogates.  The range of
non-overlong surrogates is:
     ASCII platforms                  EBCDIC I8
     "\xed\xa0\x80"               "\xf1\xb6\xa0\xa0"
to   "\xed\xbf\xbf".              "\xf1\xb7\xbf\xbf"

In both platforms, if we have the first two bytes, we can tell if it is
a surrogate or not, as all legal continuations in the rest of the byte
positions are for surrogates.  If we have only one byte, we can't tell,
so we have to assume it isn't a surrogate.

Overlongs don't meaningfully change things.  The shortest ASCII overlong
for the first surrogate is          "\xf0\x8d\xa0\x80"
and for the highest surrogate it is "\xf0\x8d\xbf\xbf".

Note that only the first byte has been changed, into two bytes.  All but
the first byte is the same for any overlong of any code point in either
ASCII or EBCDIC.

This means the algorithm for filling things out works for these two
classes in all cases.  Note also that the upper end of the range
conveniently works out without any extra effort needed.  The highest
surrogate corresponds to the highest continuation bytes.  And the
highest super that fits in the platform will also use the highest
continuation bytes.

The start bytes that need to have the fix in this commit are the ones
that could be the start of overlongs, minus the lower ones which can
represent only code points smaller than any of the ones the caller can
flag as being "interesting" (U+D800 is that value), and minus 0xFF.
Hence 0xE0 can have overlongs, but it and its overlongs can only
represent code points lower than 0xD800.  So we don't have to worry
about it or any smaller start byte.

But the reason 0xFF doesn't have to be considered is more complex.
It isn't the second byte in a sequence beginning with FF that needs to
be higher than the minimum continuation, but one further in.  This
would make things harder except that any sequence beginning with 0xFF is
Perl-extended UTF-8, and has already been considered earlier in this
function.  This code is only executed when 'must_be_super' is false.
'must_be_super' is set true if the sequence overflows or there is no
detectable overlong.  By DeMorgan's laws, this means to get here, it
doesn't overflow, and must be overlong.  To know that it is overlong, we
must have seen enough bytes to get past the point where we need a higher
continuation byte to legally fill it out.  So we can just fill the rest
with the minimum continuation.

(Note that the same reasoning would apply to 0xFE on ASCII platforms.
That is also used only by Perl-extended UTF-8, so would have been
considered earlier, and to get here we know it has to be overlong, and
so we've already seen enough bytes to not need to handle it specially.
But it fits into the same paradigm as the lower start bytes with just
the second byte needing to be higher, and there is no extra code
required to handle it besides including a case: for it in the switch().
This works in both ASCII and EBCDIC.)