utf8_to_uv_msgs: Create a common macro
Previous commits have allowed the beginning of several of the case
statements in this switch() to have the same code.  This commit creates
a macro encapsulating that code and changes the cases to use it.

The macro continues the enclosing loop if no message needs to be
generated.  This allows the removal of various conditional blocks.  And
it means that these conditions don't break to the bottom of the switch()
if no message is needed.

Braces are needed in one case: so as to not run afoul of C++
initialization crossing