ParseXS: fix SCOPE keyword and update docs
The SCOPE: keyword, which enables wrapping an XSUB's main body with
ENTER/LEAVE, has been partially broken since 5.12.0. This commit fixes
that, adds tests, and updates the very vague documentation for SCOPE in
perlxs.pod.

AFAIKT, neither the SCOPE keyword, nor it's associated /* SCOPE */
magic comment in typemap files, are used anywhere in core or on CPAN,
nor in any tests. ('SCOPE: DISABLE' appears in a single test file, but
disabled is the default anyway.)

Background:

The SCOPE keyword was added by perl-5.003_03-21-gdb3b941461 (with
documentation added soon after by perl-5.003_03-34-g84287afe68).
This made the SCOPE: keyword an XSUB-body-scoped keyword, e.g.

    void
    foo()
        SCOPE: ENABLED
        CODE:
            blah

where the emitted 'blah' code would now be wrapped with ENTER/LEAVE.

13 years later, with v5.11.0-30-g28892255e8, this was extended so that
the keyword could appear just before the XSUB too:

    SCOPE: ENABLED
    void
    foo()
        CODE:
            blah

I don't know what the motivation was behind this; the commit was part of
a larger upgrade, which just listed among other bug fixes:

    - Fix the SCOPE keyword [Goro Fuji]

but I can't find any trace of a corresponding problem description on p5p
or RT.

This change had the unfortunate side-effect of breaking the existing
XSUB-scoped variant. This is indirectly due to the fact that XSUB-scoped
KEYWORD_handler() methods are supposed to set $_ to the next line before
returning, while file scoped ones aren't supposed to.

That change made SCOPE_handler() both file- and xsub-scoped, and also
made it no longer update $_. So the new file-scoped variant worked,
while the old xsub-scope variant broke, because it now retuned with
$_ set to 'ENABLE' rather than to the next line.

The temporary fix in this commit makes SCOPE_handler() check who its
caller is and sets $_ (or not) accordingly. A proper fix will occur
shortly when a SCOPE Node subclass is added, since the NODE::parse()
methods don't pass values back and forth in $_.

This commit also updates the pod for SCOPE, which was very vague about
what the SCOPE keyword did and where it should go, syntax-wise.  I also
changed it so that it suggests the magic comment token in a typemap
entry should be /* SCOPE */. The actually regex is {/\*.*scope.*\*/}i,
which matches a whole bunch of stuff. If we ever make it stricter,
insisting on an upper-case SCOPE with just surrounding white space seems
the way to go.