pv_uni_display: Revert use of utf8_to_ov_or_die
This fixes #23089

Croaking when someone is trying to output a malformed UTF-8 sequence
turns out to be a bad idea.  That should have been obvious to me.

But instead of going back to the problematic utf8_to_uvchr_buf(), this
commit uses the non-croaking utf8_to_uv().

Croaking should not be done when at the user interface.  It makes sense,
for example, when things should have already been validated and there is
no good option to proceed.