ParseXS: process NOT_IMPLEMENTED_YET later
The undocumented and almost-entirely-unused NOT_IMPLEMENTED_YET keyword
can be used at the same point in parsing where CODE: or PPCODE: could
appear, and emits a croak() call whereas a call to a C library function
would otherwise have been auto-generated.

This keyword was checked for for partway during emitting of
initialisation code; this meant that K&R-style declarations were
emitted, but ANSI_style ones weren't.

This commit moves the checking for the presence of this keyword to a bit
later: after all initialisation code emitting is complete.

This makes NOT_IMPLEMENTED_YET logically part of the body-processing
section, which now looks roughly like

    if (/NOT_IMPLEMENTED_YET/)
        emit croak
    elsif (/PPCODE:/)
        ...
    elsif (/CODE:/)
        ...
    else
        emit autocall

and so makes the parsing code cleaner.

Conceptually it means that a NOT_IMPLEMENTED_YET can now appear after
an INIT keyword; in practice, only *INPUT* section parsing is
special-cased to recognise NOT_IMPLEMENTED_YET as another valid keyword
which terminates the current section. So INIT, C_ARGS etc sections
continue to see "NOT_IMPLEMENTED_YET" as just a bit of text to be
consumed from the input stream and added to the init code or the C
signature or whatever.  Some tests have been added to confirm this.