Clone R-O inversion list before modifying it
Unicode-defined properties are kept in inversion lists in read-only
memory.  Attempting to modify them will result in a segfault.  The new
Unicode 16.0 has some empty properties that need to be inverted before
use in some cases.  This situation has not occurred before.  The
inversion code that gets called presumes it has a modifiable SV and
inverts in-place.  This segfaults.

Normally, no inversion is needed, and the properties are used as-is, and
no problem arises.

The solution here is to create a clone, and then invert that.