ParseXS: refactor: add input_part node
Add this class:

    ExtUtils::ParseXS::Node::input_part

and give it a parse() method. This is work-in-progress; in particular
the parse() method currently also emits code, which will be moved to the
as_code() method shortly. In fact this commit mainly just converts the
block of code ripped out of the main parsing loop in process_file() in
the previous two commits into a Node subclass, mainly by doing
s/$self/$pxs/g. It also replaces the call to process_keywords() with
an inlined-version which doesn't expect $_ to be set in entry and
return. That inlined code will eventually be turned into a proper
method.

The intent of the input_part node type is to be the parent of all the
nodes created when parsing the "input" part of an XSUB - e.g. INPUT,
PREINIT, C_ARGS etc.