mktables: Handle Unicode 16.0 DoNotEmit.txt
We handle it by ignoring this file, new to Unicode 16.0.

It consists of lists of characters that, to put it less delicately than
Unicode would like, they regret creating.

But there are no rules associated with them.  It would be nice to have a
\p{DoNotEmit} property so that applications could handle situations
where this occurs.  But I'm fearful that if we did something like this,
that Unicode would later come up with something that had the same
intention but would be subtly or unsubtly different.

That has happened before, to our detriment.

So I think we should wait to see what they do do, in future releases.