mktables: Support new Unicode 16.0 properties ID_Compat_Math_foo
These new properties are automatically handled, but there is a problem.
They have no short form names.  Files are written for them based on
their names, and those files are not distinguishable on a DOS 8.3 file
system.  The solution here is to manually override the automatically
generated file names with distinguishable ones.