mktables: Handle Unicode 16.0 new \d ranges
mktables does a lot of sanity checks on the data it gets fed.  One of
those is to make sure any \d group of code points is 10 long.  This
verifies that Unicode has given us enough code points to form 0-9.  It
assumes that if it got this much right, that their numeric values are
also 0-9.  This check has uncovered issues with the Unicode Standard in
the past.

Nowadays, they've cleaned up their act, and it's been many releases
since there has been problems.  But our checks remain, and I think they
should.

What happens in Unicode 16.0 was there was a range of \d characters that
contain two consecutive groups of 0-9 values.  The check could be
changed to verify that the count is divisible by 10, but checking for
this particular range is a bit safer.