ParseXS: refactor: mark INPUT node as implicit
Give the Node::INPUT class a boolean field called 'implicit', which
signifies an INPUT section like:

    foo(a, b)
        int a
        int b

as opposed to the explicit:

    foo(a, b)
      INPUT:
        int a
        int b

Also, only add an implicit INPUT node as a kid to the input_part node if
it has kids - i.e. if there are actually any implicit input lines
present.

Neither of these changes is needed for correct functioning, but they
might be useful for a hypothetical deparser for example.