ParseXS: refactor: OUTPUT_handler() use $line vs $_
This is #1 of a small series of commits to refactor the OUTPUT_handler()
method and turn it into a Node subclass method. This series is very
similar to the one earlier in this branch which did the same for
INPUT_handler().

This commit changes the main loop from using $_ to hold the current line,
to using the variable $line instead.