ParseXS: refactor: OUTPUT_handler() split into two
This is #2 of a small series of commits to refactor the OUTPUT_handler()
method and turn it into a Node subclass method.

This commit splits the method into two: a smaller outer one which
has the 'foreach line' loop, and a new method, OUTPUT_handler_line()
which contains the bulk of the old method and processes a single line
from an OUTPUT section.