ParseXS: refactor: add Node::OUTPUT_line fields
This is #6 of a small series of commits to refactor OUTPUT keyword
handling.

The main job of parsing an OUTPUT line is to extract any information on
that line and use it to update the associated Param object (which was
likely created earlier when the XSUB's signature was parsed).

This commit makes that information also be stored in new fields in the
OUTPUT_line object. These new fields aren't currently used for anything,
but they could in principle become useful if options for deparsing or
exporting were added to ParseXS.