Nuke the MGf_PERSIST flag, freeing a bit in ->mg_flags
The MGf_PERSIST flag is only used by PERL_MAGIC_lvref, and there it
simply contains a shadow copy of the OPpLVREF_ITER flag in op_private.
Since the mg_priv field is set as a copy of op_private anyway, the code
that checks this bit might as well look there instead.

As a result of this change, the 0x80 bit of the mg_flags field is now
free for other purposes.