mk_invlists/regexec.c: Generate and use macros
With this commit, mk_invlists.pl now generates #define macros isFOO that
regexec.c now uses to determine if a character is in a particular line
breaking class.  Previously, x == foo was used.  This change insulates
the code from having to worry about when classes get changed to be
combinations.