PathTools: check Config values early before chdir
In a perl core build, the @INC paths will be relative. If you change
directories before loading all of the files needed, they won't be able
to be found. Config.pm loads some of its values at runtime, so if they
are accessed after a chdir, it may fail.

One of the PathTools tests was relying on the fact that Config_heavy.pl
would be loaded by Test::More, before it did a chdir. Newer Test::More
won't do that, so the test would fail. Accessing the required Config
values early will prevent this failure.

This same issue is unlikely to impact anything outside core, as it
requires the perl core paths in @INC to be relative paths, which is
normally not the case.