ParseXS: refactor: pass xbody to foo_part::as_code
Pass the current Node::xbody object as an extra parameter to the various
foo_part::as_code() methods. The next few commits will make use of it.

Also, in one particular loop, use the full name of the foo_part classes
for easier searching in the src code (searching for "input_part" wasn't
finding where an object of that class was being created).