debugger source lines: use PVIV instead of PVMG for source lines
These were saved as PVMG but as bulk88 suggested in
https://github.com/Perl/perl5/pull/23171#issuecomment-2780007725
we only need PVIV, since the source lines don't need magic,
aren't blessed and store an integer, not an NV.

So create them as PVIV instead.

If it turns out we do need PVMG instead for some future use, simply
remove the test added here, it's simply to confirm we don't need
PVMG here.