mktables: Consolidate code into a single function
Some properties in Unicode essentially form equivalence classes for all
possible code points.

For example, Unicode publishes the Line Break (LB) property, where each
possible code point is given a type, like Alphabetic, or Opening
Parenthesis.  All code points that act as alphabetics have the AL
equivalence class.  All that act like Opening Parentheses have the OP
class.

Unicode also publishes rules as to if it is permissible to break between
code point of any types.  For the Line Break property, you wouldn't
break a line between two alphabetics or between an opening parenthesis
and an alphabetic, but you could between a Space and almost any other
type or between a closing parenthesis and many types.

Perl uses these properties to implement the \b{lb} etc regular
expression constructs.  It uses a two-dimensional array where the value
in the cell [x,y] tells whether a break is permissible between
characters of type x and characters of type y.  (Some cases can't be
done with this simple lookup, but knowing the surrounding context is
necessary to make a decision.  Those are implemented as DFAs in
regexec.c.)

Unicode used to publish such an array for the Line Break property, and
still publishes some non-normative .html files that contain similar
information.  But to really know what to do, one has to read documents
UAX#14 and UAX#29 that contain textual descriptions of the rules.  These
change each new release, and are the major pain in upgrading to a new
release.

In recent releases, Unicode has mostly stopped creating new equivalence
classes as it has refined the rules for the boundary conditions  For
example, the line boundary conditions are very different for East Asian
(EA) characters than the Western scripts.  Effectively there are thus
two sets of rules.  But instead of creating new equivalence classes that
reflect this reality, Unicode has chosen to just document it in those
two UAX documents.  I don't know the motivation for this.

But perl wants that table to divvy up all the possible boundary
conditions, so it can continue to use the array to make most of the
decisions, so mktables splits the equivalence classes that Unicode
provides into new ones that reflect what the UAXes say.  At first, I
thought this was a one-off matter, so wrote a few lines to handle a
special case; then when the next release came out, added a few more for
another one, etc.  But Unicode 15.1 and 16.0 continue the trend, so it's
become an effort.

This commit consolidates the previous one-off code snippets into one
generalized function.  It should be able to handle future instances
without having to craft something new each time.

It also creates a new data structure that mk_invlists.pl can look at so
that it doesn't have to repeat the logic found here, as it currently
does.