mk_invlists: Use new set subtraction ability
This allows the removal of some combinatorial complexity, thus showing
a bug in which the combination of PO to EOP had not been added when it
should have been.

Currently, mktables splits the Line Break OP and CP classes into East
Asian ones, and the remainders.  The extra combinations occurred because
the code here needed to take every existing OP and add an East_Asian
(EA_OP) equivalent; same with CP.  It's easy to miss one, and I did.

This commit allows this split to be hidden from most places in
mk_invlists.