mk_invlists: Move break tables edge cells
These tables are placed in charclass_invlists.h.  They have a row and
column for what happens when the position being checked for is at the
start or end of the text.  This commit reorders the tables so that the
edge row and column are, well, at the edges.  And it relabels the labels
to be '^' and '$' respectively.