ParseXS: refactor: rm xsub_targ_declared_early
Remove the field 'xsub_targ_declared_early' from the ParseXS class,
and instead add a 'use_early_targ' field to the ReturnType Node class.

This is the first in a series of commits which will remove many of the
'global' parse state fields from the ParseXS object and instead add them
as fields of the relevant Node objects.