ParseXS: refactor: remove xsub_seen_RETVAL_in_CODE
Remove the following field from the ExtUtils::ParseXS class:

    xsub_seen_RETVAL_in_CODE

and replace it with this new field in the ExtUtils::ParseXS::Node::xbody
class:

    seen_RETVAL_in_CODE

Note that the previous commit did similar things, but for fields added
to the Node::xsub class. This flag is per-CASE, so goes in Node::xbody
instead. Also add a test for multiple CASEs.