mk_invlists: Use new split capability with ExtPict
\p{Extended_Pictographic} is not fully implemented yet because unlike
other properties, it can match a string instead of a single character.

And it is kind of a kludge here  The 14.0 release was analyzed by me and
the rules here were customized based on that analysis.  For example, in
the Line Break property, a clause was added by Unicode to Rule LB30b
that required taking the intersection of this property and all the
Unassigned code points.  It turns out that everything in that
intersection had the Line Break class of Ideographic, so I modified
mktables to split the Ideographic class into two components, the
elements of the intersection went into the long-named
"Unassigned_Extended_Pictographic_Ideographic" and plain Ideographic was
left with the remainder.  To match all of Ideographic you have to
specify both classes.  By using the new split capability, this can be
done effectively as a macro expansion, and the special cases can be
removed from the code.  This commit does this.

Similarly, both the Word Break and Grapheme Cluster Break properties
have somewhat different interactions with Extended_Pictographic that
this commit smooths over.

This situation is brittle.  A new release of Unicode might change things
so that Ideographic isn't the only LB class in the intersection
mentioned above, so the customization has to be checked in every
release.  A few commits later in this branch, this will be automated,
and no longer a concern.