ParseXS: refactor: rm xsub_implicit_OUTPUT_RETVAL
Remove the following field from the ExtUtils::ParseXS class:

    xsub_implicit_OUTPUT_RETVAL

and replace it with this new field in the ExtUtils::ParseXS::Node::xbody
class:

    seen_autocall

Basically there is some logic which says that a bodiless (i.e. autocall)
XSUB has an implicit 'OUTPUT: RETVAL' - i.e. return RETVAL even though
it hasn't explicitly been requested. This commit moves the logic
to the point where is RETVAL is emitted, and earlier just notes that an
autocall is present.

This commit also fixes a bug introduced by the earlier refactoring in
this branch which split out parsing and code emitting. When an XSUB has
two or more CASEs, the xbody's weren't treated independently - so the
behaviour of all CASE branches was determined by the state of the last
branch. With seen_autocall being set per xbody, the corrrect behaviour is
now restored. This commit adds a couple of tests to confirm this.