mk_invlists: Reverse order of break property rules
Before this commit, the rules for populating the tables for break
properties were laid out in reverse order, so that the lowest priority
rule was executed first.  It filled a cell, which then would be
overwritten by any higher priority rule that applied to it.

This reverse order made it harder to compare the rules with the text of
the Unicode rules these are trying to implement.

This commit changes things to have the rules in the same order as
Unicode lists them.

The previous scheme had certain advantages that this has to make up for
by using temporary code to override what would otherwise have gone into
the cells.  This code will no longer be needed in a few commits when a
general purpose stacking DFA scheme is implemented.

As a result, of this temporary code, only two cells in one property
change as a result of this complete reversal.  They change to using a
DFA which ends up returning the same results as the original
unconditional value.