mk_invlists: Remove some special cases
These were added to compensate for reversing the order of handling the
break property rules.  This commit hides the need for that in one place
per table, except for a second place for Line Break.

The only changes to the tables occur in the garbage row and column which
aren't actually accessed, so those changes are harmless.

It is a temporary commit.  A few commits from now, this will be removed.