mk_invlists: Generalize to stack DFAs for break properties
The Unicode breaking algorithms are supposed to be implemented by
executing DFAs in priority order, stopping at the first one that
succeeds.  (In many cases a DFA isn't needed, and we can unconditionally
say that there is or isn't a break at a given position simply by looking
at the characters on either side of it.)

But it was a significant amount of work to get from where perl started
to be able to do that.  And it hasn't been necessary until now.  In most
cases, a single DFA suffices, and where not, a more complicated single
DFA took care of the stacking.

But this has become untenable in Unicode 15.1, so I ended up doing the
work to implement their algorithm.  The result is more, but simpler
DFAs, and it becomes easier to add new ones, as they don't have to
interact with other ones.  The stacking does that for them.

This commit implements a separate DFA table beyond the x,y lookup table.
If the decision that this is a breakable position requires a DFA, the
x,y contents are an index into this separate table, which contains the
DFA to follow.  The first element gives the case statement number to use
to execute the DFA.  The second element gives the value to return if the
DFA succeeds.  If it fails, the code add +2 to get the next thing to
try.