ParseXS: refactor: delete param field from Params
The Node::Params class has a 'params' field which holds a list of
Node::Param objects. This class was one of the first Node classes to be
created during my recent refactoring work, and at the time, Node
subclasses didn't have a generic 'kids' field. They do now, so just
store the list of Param objects of 'kids' of the Params object.