ParseXS: refactor: set default alias at parse time
If the list of aliases for an XSUB doesn't include the XSUB's main name,
an extra alias entry is added, mapping the main name to ix 0.

Move this setting from the code generation phase to the end of the
parsing phase, because the AST should really be complete by the end of
parsing.

Also add a test for this behaviour.

Shouldn't affect hat code is generated.