ParseXS: refactor: add Param->parse() method.
This commit just moves a block of code of the form sub {...}->()
into its own named sub. There are no changes to the moved lines of code
apart from indentation.

This is the second of three commits to create the parse() method.
The next commit will do any final tidying up.