mk_invlists: Use new split capability with ALetter
ALetter also contains the class ExtPict_LE.  Prior to this commit, there
had to be a rule for each ALetter doing the same thing with ExtPict_LE.
But the new splits capability allows ALetter to expand automatically to
both.

This uncovers a bug.  There should have been a rule
 WB5 ALetter x ExtPict_LE
which was missing.