regexec.c: Skip CM and ZWJ in look behind in LB parsing
The Unicode standard says that these two characters are to be ignored
for the purposes of determining if there is a Line Break just before
certain characters.  That is, you have to back up in the parse string
past all adjacent ones of these, and then examine it.

This applies to any lower priority rule than LB9.  This commit fixes two
cases that didn't do that.