mk_invlists: Use new DFA scheme for horizontal white space
Perl doesn't follow the Unicode standard with regard to its treatment of
white space, in particular sequences of horizontal white space.  Unicode
allows "tailoring" of its rules for local situations, and Perl
traditionally with \B  has treated all sequences of white space as a
single unit.  Unicode originally considered each space in a sequence of
them as a separate unit.  A perl program would want them all a single
unit.  Unicode eventually came round to our way of thinking, but not
entirely, as comments unaffected by this commit indicate.

The DFA for this situation does not fit in with the new stackable DFA
scheme, and woul start failing tests a few commits later as the shim
code is removed.  Convert to the new scheme, which allows us to call the
functions that affect a single cell twice with effect.  The order is
immaterial, but one call installs a default behavior, and the other a DFA
that ends up being executed first to override that behavior in certain
(rare) situations.